Move some compute helpers separately.
